+++ perfkitbenchmarker/providers/aws/aws_network.py
@@ -38,6 +38,8 @@ from perfkitbenchmarker import resource
 from perfkitbenchmarker import vm_util
 from perfkitbenchmarker.providers.aws import aws_placement_group
 from perfkitbenchmarker.providers.aws import aws_vpc_endpoint
+from perfkitbenchmarker.providers.aws import aws_elastic_ip
+from perfkitbenchmarker.providers.aws import aws_global_accelerator
 from perfkitbenchmarker.providers.aws import util
 
 _AWS_VPC = flags.DEFINE_string(
@@ -849,8 +851,8 @@ class AwsNetwork(network.BaseNetwork):
     self.global_accelerator = None
     if FLAGS.aws_global_accelerator:
       logging.warn("using aws global accelerator")
-      self.elastic_ip = AwsElasticIP(self.region)
-      self.global_accelerator = AwsGlobalAccelerator()
+      self.elastic_ip = aws_elastic_ip.AwsElasticIP(self.region)
+      self.global_accelerator = aws_global_accelerator.AwsGlobalAccelerator()
 
   @staticmethod
   def _GetNetworkSpecFromVm(vm):
@@ -876,19 +878,18 @@ class AwsNetwork(network.BaseNetwork):
       self.placement_group.Create()
     if FLAGS.aws_global_accelerator:
       logging.warn("using aws global accelerator")
-      self.elastic_ip._Create()
-      self.global_accelerator._Create()
+      self.elastic_ip.Create()
+      self.global_accelerator.Create()
       self.global_accelerator.AddListener('TCP', '10', '60000')
       self.global_accelerator.listeners[-1].AddEndpointGroup(
                                             self.region,
                                             self.elastic_ip.allocation_id, 128)
 
-
   def Delete(self):
     """Deletes the network."""
     if FLAGS.aws_global_accelerator:
-      self.global_accelerator._Delete()
-      self.elastic_ip._Delete()
+      self.global_accelerator.Delete()
+      self.elastic_ip.Delete()
     if self.subnet:
       self.subnet.Delete()
     if self.placement_group:
@@ -976,435 +977,3 @@ class AwsVpcPeering(network.BaseVPCPeering):
         '--region=%s' % self.network_a.region,
         '--vpc-peering-connection-id=%s' % self.id]
     vm_util.IssueCommand(delete_cmd)
-
-
-
-
-class AwsGlobalAccelerator(resource.BaseResource):
-  """An object representing an Aws Global Accelerator.
-  https://docs.aws.amazon.com/global-accelerator/latest/dg/getting-started.html"""
-
-# {
-#    "Accelerator": { 
-#       "AcceleratorArn": "string",
-#       "CreatedTime": number,
-#       "Enabled": boolean,
-#       "IpAddressType": "string",
-#       "IpSets": [ 
-#          { 
-#             "IpAddresses": [ "string" ],
-#             "IpFamily": "string"
-#          }
-#       ],
-#       "LastModifiedTime": number,
-#       "Name": "string",
-#       "Status": "string"
-#    }
-# }
-
-  def __init__(self):
-    super(AwsGlobalAccelerator, self).__init__()
-    # all global accelerators must be located in us-west-2
-    self.region = 'us-west-2'
-    self.idempotency_token = None
-
-    #The name can have a maximum of 32 characters, 
-    #must contain only alphanumeric characters or hyphens (-), 
-    #and must not begin or end with a hyphen.
-    self.name = None
-    self.accelerator_arn = None
-    self.enabled = False
-    self.ip_addresses = []
-    self.listeners = []
-
-# aws globalaccelerator create-accelerator 
-#         --name ExampleAccelerator
-#         --region us-west-2
-#         --idempotencytoken dcba4321-dcba-4321-dcba-dcba4321
-
-  def _Create(self):
-    """Creates the internet gateway."""
-    if not self.idempotency_token:
-      self.idempotency_token = str(uuid.uuid4())[-50:]
-
-    self.name = 'pkb-ga-%s-%s' % (FLAGS.run_uri, str(uuid.uuid4())[-12:])
-
-    create_cmd = util.AWS_PREFIX + [
-        'globalaccelerator',
-        'create-accelerator',
-        '--name', self.name,
-        '--region', self.region,
-        '--idempotency-token', self.idempotency_token]
-    stdout, _, _ = vm_util.IssueCommand(create_cmd)
-    response = json.loads(stdout)
-    self.accelerator_arn = response['Accelerator']['AcceleratorArn']
-    self.ip_addresses = response['Accelerator']['IpSets'][0]['IpAddresses']
-    logging.info("ACCELERATOR IP ADDRESSES")
-    logging.info(self.ip_addresses)
-    #util.AddDefaultTags(self.id, self.region)
-
-  #@vm_util.Retry()
-  def _Delete(self):
-    """Deletes the Accelerator."""
-
-    # need to disable accelerator before it can be deleted
-    self._Update(enabled=False)
-    status = self.Describe()
-    while status['Accelerator']['Enabled'] == True:
-      status = self.Describe()
-
-    # need to delete listeners before accelerator can be deleted
-    for listener in self.listeners:
-      listener._Delete()
-
-    delete_cmd = util.AWS_PREFIX + [
-        'globalaccelerator',
-        'delete-accelerator',
-        '--region', self.region,
-        '--accelerator-arn', self.accelerator_arn]
-    output = vm_util.IssueRetryableCommand(delete_cmd)
-
-    exists = self._Exists()
-    while exists:
-      stdout, stderr, _ = vm_util.IssueCommand(delete_cmd, raise_on_failure=False)
-      if "AcceleratorNotFoundException" in stderr:
-        break
-      exists = self._Exists()
-
-  #@vm_util.Retry()
-  def _Update(self,enabled):
-    """Returns true if the internet gateway exists."""
-    update_cmd = util.AWS_PREFIX + [
-        'globalaccelerator',
-        'update-accelerator',
-        '--region', self.region,
-        '--accelerator-arn', self.accelerator_arn]
-    if enabled:
-      update_cmd += ['--enabled']
-    else:
-      update_cmd += ['--no-enabled']
-    stdout, _ = util.IssueRetryableCommand(update_cmd)
-    response = json.loads(stdout)
-    accelerator = response['Accelerator']
-    # assert accelerator['Enabled'] == enabled, 'Accelerator not updated'
-
-  def _Exists(self):
-    """Returns true if the accelerator exists."""
-    describe_cmd = util.AWS_PREFIX + [
-        'globalaccelerator',
-        'describe-accelerator',
-        '--region', self.region,
-        '--accelerator-arn', self.accelerator_arn]
-    try:
-      stdout, _, _ = vm_util.IssueCommand(describe_cmd, raise_on_failure=False)
-      response = json.loads(stdout)
-      accelerator = response['Accelerator']
-      return len(accelerator) > 0
-    except ValueError as e:
-      return False
-
-
-
-  def Describe(self):
-    """Returns true if the accelerator exists."""
-    describe_cmd = util.AWS_PREFIX + [
-        'globalaccelerator',
-        'describe-accelerator',
-        '--region', self.region,
-        '--accelerator-arn', self.accelerator_arn]
-    stdout, _ = util.IssueRetryableCommand(describe_cmd)
-    response = json.loads(stdout)
-    return response
-
-  def Status(self):
-    """Returns true if the accelerator exists."""
-    describe_cmd = util.AWS_PREFIX + [
-        'globalaccelerator',
-        'describe-accelerator',
-        '--region', self.region,
-        '--accelerator-arn', self.accelerator_arn]
-    stdout, _ = util.IssueRetryableCommand(describe_cmd)
-    response = json.loads(stdout)
-    status = response['Accelerator']['Status']
-    return status
-
-  #  @vm_util.Retry(poll_interval=1, log_errors=False, max_retries=5
-  #                retryable_exceptions=(AwsTransitionalVmRetryableError,))
-  def isUp(self):
-    """Returns true if the accelerator exists."""
-    describe_cmd = util.AWS_PREFIX + [
-        'globalaccelerator',
-        'describe-accelerator',
-        '--region', self.region,
-        '--accelerator-arn', self.accelerator_arn]
-    stdout, _ = util.IssueRetryableCommand(describe_cmd)
-    response = json.loads(stdout)
-    status = response['Accelerator']['Status']
-    return status
-
-  def AddListener(self, protocol, start_port, end_port):
-    """Returns true if the accelerator exists."""   
-    self.listeners.append(AwsGlobalAcceleratorListener(self,
-                                                       protocol,
-                                                       start_port,
-                                                       end_port))
-    self.listeners[-1]._Create()
-
-
-class AwsGlobalAcceleratorListener(resource.BaseResource):
-  """Class representing an AWS Global Accelerator listener."""
-
-  def __init__(self, accelerator, protocol, start_port, end_port):
-    super(AwsGlobalAcceleratorListener, self).__init__()
-    self.accelerator_arn = accelerator.accelerator_arn
-    #self.target_group_arn = target_group.arn
-    self.start_port = start_port
-    self.end_port = end_port
-    self.protocol = protocol
-    self.region = accelerator.region
-    self.idempotency_token = None
-    self.arn = None
-    self.endpoint_groups = []
-
-# aws globalaccelerator create-listener 
-#        --accelerator-arn arn:aws:globalaccelerator::012345678901:accelerator/1234abcd-abcd-1234-abcd-1234abcdefgh 
-#        --port-ranges FromPort=80,ToPort=80 FromPort=81,ToPort=81 
-#        --protocol TCP
-#        --region us-west-2
-#        --idempotencytoken dcba4321-dcba-4321-dcba-dcba4321
-
-  def _Create(self):
-    if not self.idempotency_token:
-      self.idempotency_token = str(uuid.uuid4())[-50:]
-    """Create the listener."""
-    create_cmd = util.AWS_PREFIX + [
-        'globalaccelerator',
-        'create-listener',
-        '--accelerator-arn', self.accelerator_arn,
-        '--region', self.region,
-        '--protocol', self.protocol,
-        '--port-ranges', 
-        'FromPort=%s,ToPort=%s' % (str(self.start_port), str(self.end_port)),
-        '--idempotency-token', self.idempotency_token
-    ]
-    stdout, _, _ = vm_util.IssueCommand(create_cmd)
-    response = json.loads(stdout)
-    self.listener_arn = response['Listener']['ListenerArn']
-    logging.info("LISTENER ARN")
-    logging.info(self.listener_arn)
-
-# RESPONSE
-# {
-#    "Listener": { 
-#       "ClientAffinity": "string",
-#       "ListenerArn": "string",
-#       "PortRanges": [ 
-#          { 
-#             "FromPort": number,
-#             "ToPort": number
-#          }
-#       ],
-#       "Protocol": "string"
-#    }
-# }
-
-  def _Exists(self):
-    """Returns true if the accelerator exists."""
-    describe_cmd = util.AWS_PREFIX + [
-        'globalaccelerator',
-        'describe-listener',
-        '--region', self.region,
-        '--listener-arn', self.listener_arn]
-    stdout, _ = util.IssueRetryableCommand(describe_cmd)
-    response = json.loads(stdout)
-    accelerator = response['Listener']
-    return len(accelerator) > 0
-
-  def _Delete(self):
-    """Deletes Listeners"""
-    for endpoint_group in self.endpoint_groups:
-      endpoint_group._Delete()
-    delete_cmd = util.AWS_PREFIX + [
-        'globalaccelerator',
-        'delete-listener',
-        '--region', self.region,
-        '--listener-arn', self.listener_arn]
-    vm_util.IssueCommand(delete_cmd)
-
-  def AddEndpointGroup(self, region, endpoint, weight):
-    """Add end point group to listener."""   
-    self.endpoint_groups.append(AwsEndpointGroup(self, region))
-    self.endpoint_groups[-1]._Create(endpoint, weight)
-
-class AwsEndpointGroup(resource.BaseResource):
-  """An object representing an Aws Global Accelerator.
-  https://docs.aws.amazon.com/global-accelerator/latest/dg/getting-started.html"""
-
-# {
-#    "EndpointGroup": { 
-#       "EndpointDescriptions": [ 
-#          { 
-#             "EndpointId": "string",
-#             "HealthReason": "string",
-#             "HealthState": "string",
-#             "Weight": number
-#          }
-#       ],
-#       "EndpointGroupArn": "string",
-#       "EndpointGroupRegion": "string",
-#       "HealthCheckIntervalSeconds": number,
-#       "HealthCheckPath": "string",
-#       "HealthCheckPort": number,
-#       "HealthCheckProtocol": "string",
-#       "ThresholdCount": number,
-#       "TrafficDialPercentage": number
-#    }
-# }
-
-  def __init__(self, listener, endpoint_group_region):
-    super(AwsEndpointGroup, self).__init__()
-    # all global accelerators must be located in us-west-2
-    self.region = 'us-west-2'
-    self.idempotency_token = None
-    self.listener_arn = listener.listener_arn
-    self.endpoint_group_region = endpoint_group_region
-    self.endpoint_group_arn = None
-    self.endpoints = []
-
-# aws globalaccelerator create-endpoint-group 
-#            --listener-arn arn:aws:globalaccelerator::012345678901:accelerator/1234abcd-abcd-1234-abcd-1234abcdefgh/listener/0123vxyz 
-#            --endpoint-group-region us-east-1 
-#            --endpoint-configurations EndpointId=eipalloc-eip01234567890abc,Weight=128
-#            --region us-west-2
-#            --idempotencytoken dcba4321-dcba-4321-dcba-dcba4321
-
-  def _Create(self, endpoint, weight=128):
-    """Creates the internet gateway."""
-    if not self.idempotency_token:
-      self.idempotency_token = str(uuid.uuid4())[-50:]
-
-    create_cmd = util.AWS_PREFIX + [
-        'globalaccelerator',
-        'create-endpoint-group',
-        '--listener-arn', self.listener_arn,
-        '--endpoint-group-region', self.endpoint_group_region,
-        '--region', self.region,
-        '--idempotency-token', self.idempotency_token,
-        '--endpoint-configurations',
-        'EndpointId=%s,Weight=%s' % (endpoint, str(weight))]
-    stdout, _, _ = vm_util.IssueCommand(create_cmd)
-    response = json.loads(stdout)
-    self.endpoint_group_arn = response['EndpointGroup']['EndpointGroupArn']
-    self.endpoints.append(endpoint)
-    #util.AddDefaultTags(self.id, self.region)
-    return
-
-  def _Update(self, endpoint, weight=128):
-    """Creates the internet gateway."""
-    if not self.idempotency_token:
-      self.idempotency_token = ''.join(
-        random.choice(string.ascii_lowercase + 
-                      string.ascii_uppercase +  
-                      string.digits) for i in range(50))
-
-    create_cmd = util.AWS_PREFIX + [
-        'globalaccelerator',
-        'update-endpoint-group',
-        '--region', self.region,
-        '--endpoint-configurations',
-        'EndpointId=%s,Weight=%s' % (endpoint, str(weight))]
-    vm_util.IssueCommand(create_cmd)
-    #util.AddDefaultTags(self.id, self.region)
-
-  def _Delete(self):
-    """Deletes the endpoint group."""
-    delete_cmd = util.AWS_PREFIX + [
-        'globalaccelerator',
-        'delete-endpoint-group',
-        '--region', self.region,
-        '--endpoint-group-arn', self.endpoint_group_arn]
-    vm_util.IssueCommand(delete_cmd)
-
-  def _Exists(self):
-    """Returns true if the internet gateway exists."""
-    describe_cmd = util.AWS_PREFIX + [
-        'globalaccelerator',
-        'describe-endpoint-group',
-        '--region', self.region,
-        '--endpoint-group-arn', self.endpoint_group_arn]
-    stdout, _ = util.IssueRetryableCommand(describe_cmd)
-    response = json.loads(stdout)
-    internet_gateways = response['EndpointGroup']
-    return len(internet_gateways) > 0
-
-
-class AwsElasticIP(resource.BaseResource):
-  """An object representing an AWS Elastic IP address"""
-
-  def __init__(self, region, domain='vpc'):
-    super(AwsElasticIP, self).__init__()
-    assert (domain in ('vpc', 'standard')), "Elastic IP domain type, %s, must be either vpc or standard" % domain
-    self.domain = domain
-    self.public_ip = None
-    self.region = region
-    self.allocation_id = None
-    self.attached = False
-    self.instance_id = None
-
-  def _Create(self):
-    """Creates the elastic IP address."""
-    create_cmd = util.AWS_PREFIX + [
-        'ec2',
-        'allocate-address',
-        '--domain', self.domain,
-        '--region', self.region]
-    stdout, _, _ = vm_util.IssueCommand(create_cmd)
-    response = json.loads(stdout)
-    self.allocation_id = response['AllocationId']
-    self.public_ip = response['PublicIp']
-    #util.AddDefaultTags(self.id, self.region)
-
-  def _Delete(self):
-    """Deletes the elastic IP."""
-    delete_cmd = util.AWS_PREFIX + [
-        'ec2',
-        'release-address',
-        '--region', self.region,
-        '--allocation-id', self.allocation_id]
-    vm_util.IssueCommand(delete_cmd)
-
-  def _Exists(self):
-    """Returns true if the elastic IP exists."""
-    describe_cmd = util.AWS_PREFIX + [
-        'ec2',
-        'describe-addresses',
-        '--region', self.region,
-        '--allocation-ids', self.allocation_id]
-    stdout, _ = util.IssueRetryableCommand(describe_cmd)
-    response = json.loads(stdout)
-    addresses = response['Addresses']
-    return len(addresses) > 0
-
-  def AssociateAddress(self, instance_id):
-    """Associates elastic IP with an EC2 instance in a VPC"""
-    if not self.attached:
-      self.instance_id = instance_id
-      attach_cmd = util.AWS_PREFIX + [
-          'ec2',
-          'associate-address',
-          '--region', self.region,
-          '--instance-id', self.instance_id,
-          '--allocation-id', self.allocation_id]
-      util.IssueRetryableCommand(attach_cmd)
-      self.attached = True
-
-  def DisassociateAddress(self):
-    """Disssociates elastic from EC2 Instance"""
-    if self.attached:
-      detach_cmd = util.AWS_PREFIX + [
-          'ec2',
-          'disassociate-address',
-          '--region', self.region,
-          '--association-id', self.allocation_id]
-      util.IssueRetryableCommand(detach_cmd)
-      self.attached = False
